Resolve path for `default_notes_file` to support symlinks.
This fixes an issue where you would end up with two references to the
default file when using symlinks, since all other files had their
symlinks resolved but not this. This led to having one reference to the
full path of the file, and one reference to the symlink, giving symptoms
such as duplicate entries in the agenda.