Improve refiling error handling
Refiling to a correct org-file with an incorrect headline resulted in an error,
due to the function _refile_to_end being called with '.' instead of ':'. Now
instead an informative error message is displayed if the user either tried to
refile to a non-existent file or to a non-existant headline in an existing file.