Force-disable spellcheck in certain markup groups.
This passes the `spell` option to `nvim_buf_set_extmark()` for all
markup groups. In most cases, the passed value is simply `nil`, i.e. the
same as if it were not passed at all.

In the following cases, however, we pass `spell=false`, i.e. forcefully
disable spell-checking:

- the URL part of `[[URL][hyperlinks]]`;
- the inside of `[[hyperlinks]]` without description;
- `~code~` and `=verbatim=` spans;
- in-line LaTeX markup using `\(parentheses\)`.

The reason in all cases is that these spans are extremely likely to
contain purposefully non-orthographic spelling, in particular
hyperlinks.