Make indentexpr() slightly easier to debug.
This simply turns the indentation line number and the current mode into
optional arguments. If they're not passed, `vim.v.lnum` and
`vim.fn.mode()` are used as usual. However, this allows calling a
*hyptothetical* `indentexpr()` from the command line for debugging.

Note that `indentexpr()` currently doesn't *use* `vim.fn.mode()`. It
will in a future commit, though.