fix(hyperlinks): support "file:../" and "../"
Following a relative link starting with "../" is now supported.

Examples:
file:../somedir/some.org
../somedir/some.org

Previously it was possible to follow the first link, but a file "false"
got opened instead of the target file. The second type was unsupported.

Because autocompletion works, it was easy for the user to create such
links and surprising, that they didn't work.