feat: allow TS users to highlight marker delimiters
This permits those of us who use treesitter to change the highlights
surrounding markers. E.g. for the "verbatim" marker we can now highlight
the "=" characters individually instead of the entire chunk being the
same highlight.

This is particularly useful in tables with lots of code elements, this
makes it easier to tell the delimiter characters apart from the primary
marker.