fix: promote headline while refile to file-end
- promote headlines to first level when archiving (like Emacs)
- promote headlines to first level when refiling without targeting a
  headline (Emacs only allows headlines as refiling targets)

Motivation:

Refiling to file-ends is only indirectly used in Emacs Org-Mode for
archiving. In this case Org-Mode promotes headlines to the first
level in the archive file.

This fix aligns the archiving-behavior with Emacs while also adapting
the normal refiling feature when no headline was given.

This has two advantages: It prevents the user from getting wrong and
unexpected structures in org-files and archive-files and it aligns
Nvim-Orgmode more with the original Emacs Org-Mode.