chore: remove superfluous refile functions
Because the main logic is now consolidated in Capture:_refile_to,
the two functions Capture:refile_to_headline and Capture:_refile_to_end
are not needed anymore.

Adjust the tests accordingly.

Note:
A general refactoring of the Capture class would be helpful in the
future. The current Capture class mixes several concerns: Refiling,
capturing and archiving. Separating these into different classes might
help to make the code easier to understand and to adjust.