feat: add virtual indentation
Allows Virtual indentation to be used simmilar to emacs
`org-indent-mode` by adding virtual spaces before text under headlines
to align them with a given headline.

This adds the following new options mirroring Emacs:
- `org_startup_indented`
- `org_adapt_indentation`
- `org_indent_mode_turns_off_org_adapt_indentation`

And deprecates the config option `org_indent_mode`.

A shim was added for backwards compatibility so this is not a breaking
change.

See `DOCS.md` for additional details.