fix(agenda): Reload agenda if headline range was changed
When completing a todo with a repeater date, date is added to the logbook.
This increases the content of the headline and shifts all headlines below the updated headline.
This movement can cause those shifted headlines to have outdated position in the agenda.
To solve this, we reload the agenda anyway and do update in place if we are able to find the headline.