fix(virtual indent): make TS lookup not exclude lines
Prior to this commit, it was possible that virtual indents would only
apply to the first heading in a range of lines and the other headings
would lose virtual indentation.

This change makes the Treesitter lookup prefer the largest range of
lines to set virtual indentation on and thus ensures virtual indent is
set correctly across the entire range.