fix: block content alignment (#698)
* fix(indent): do not modify block indents if one line is correct

Previously, if a single line was correctly aligned with a block, but the
others were overindented it was possible for the overindented lines to
be unindented. This is not correct.

The overindented lines shouldn't shift at all if at least one of the
lines is correctly indented. This is to preserve manual formatting.

This commit ensures the above functionality.

* fix(indent): ignore blank lines for block content indentation