docs(README): add custom_ids and fix relative links
In nvim-orgmode, you can point to a headline either by using the star
syntax (* HEADLINE) or by using a custom-id in the properties drawer of
the headline, and then refer to it using the hash syntax (#custom-id).

This commit adds custom_ids for all the headlines referred to into the
README.org file and fixes relative links to other files in the
repository.