fix: generate missing intermediate priorities
A headline priority gets only detected correctly, if it is part of the
table returned by Config:get_priorities(). While it previously only
returned the priorities explicitly defined by the user, it now uses
PriorityState to generate also intermediate ones. This happens, if the
highest and the lowest priority define a range of more than three
priorities.