Fix priorities (#784) (#785)
* test: implement test to expose the bug

* test: add more cases in priority_state_spec

* refactor: make priority_state independent of config

To prepare the fix, we need to avoid a dependency cycle between config
and priority_state, because we will need priority_state in config to
implement the fix.

We inject the needed values, whereever an instance of PriorityState is
created.

* fix: generate missing intermediate priorities

A headline priority gets only detected correctly, if it is part of the
table returned by Config:get_priorities(). While it previously only
returned the priorities explicitly defined by the user, it now uses
PriorityState to generate also intermediate ones. This happens, if the
highest and the lowest priority define a range of more than three
priorities.

* docs: mention new highlight groups

Add new highlight groups to "Colors" chapter.

---------

Co-authored-by: Sebastian Flügge <seflue@users.noreply.github.com>