Limit get_date_under_cursor to parsing current line
To retrieve a date under the cursor, limit the scope to the current
line. Parse all dates on the line, then filter the table to retrieve
the member containing the cursor.

Addresses the TODO that would fail to return the correct entry when
multiple are found on a given line.