fix(meta_return)!: insert new line item below cursor
This is a fix for org-meta-return to make it work in the same way as
Emacs orgmode for line items.
Previously, new line items were added after the current line item, and
all of it's children content (nested line items).
Now, new line item is added right below the current line item, pushing
it's child down.