feat(attach): add `OrgAttach:attach_to_other_buffer()`
This function has no direct Emacs equivalent. However, it tries to
replicate the functionality provided by `org-attach-dired-to-subtree`,
which attaches files selected in a dired window (roughly equivalent to
Vim's NetRW) to the last-seen Orgmode task.

Our implementation tries to be extremely general because the Neovim
ecosystem has a plethora of file browser plugins.