feat(mappings): open at point opens files
Suppose you want to link to a PDF file on your system. If the (relative)
file path does not include spaces, you can use `gx` mapping from neovim.
However, when we have a nice url handler in orgmode.nvim, we can
leverage that and pass the parsed url to `vim.ui.open` (same function
that `gx` calls).
Currently, only urls that start with `/` or `~` or `.` are considered
"files".