feat(indent)!: Add `Org indent_mode` command in favor of toggling `vim.b.org_indent_mode` (#932)
Previously, toggling virtual indentation on the buffer was done by
manually changing value of the buffer variable `vim.b.org_indent_mode`
to true or false respectively.
This PR introduces an subcommand `indent_mode` to the global `Org`
command which will toggle the virtual indentation on/off.
This allows us to simplify and remove some of the code that might become
deprecated in the future.