feat(todos): support file-specific todo definitions
Support to overwrite the globally defined todo keywords with a file
specific one.

While emacs orgmode actually supports multiple todo keyword sequences,
this implementation explicitly focuses on one sequence, because the
plugin currently also supports only one sequence to be defined in the
configuration.

To support the full emacs functionality this needs to be extended later.