feat(todos): support file-specific todo definitions (#956)
* feat(todos): support file-specific todo definitions

Support to overwrite the globally defined todo keywords with a file
specific one.

While emacs orgmode actually supports multiple todo keyword sequences,
this implementation explicitly focuses on one sequence, because the
plugin currently also supports only one sequence to be defined in the
configuration.

To support the full emacs functionality this needs to be extended later.

* fix: meomize parsed todo keywords

Co-authored-by: Kristijan Husak <husakkristijan@gmail.com>

* fix: trim spaces

Co-authored-by: Kristijan Husak <husakkristijan@gmail.com>

* fix: support file specific todos in all todo-related commands

* test: add permutation test for file specific todos

---------

Co-authored-by: Sebastian Flügge <sebastian.fluegge@dnv.com>
Co-authored-by: Kristijan Husak <husakkristijan@gmail.com>