feat(agenda)!: Add time grid
This is considered a breaking change commit because it adds a new time
grid option to the agenda view, which was not there before.
It is set to `true` by default to align with how Emacs Orgmode works out
of the box.

To disable the time grid, set `org_agenda_use_time_grid = false` in your
orgmode config.