fix(windows): handle Windows file paths correctly in refile
There was a bug on Windows where the first character of the file name was dropped during refile. This happened because an extra leading / was added to the beginning of the path, resulting in something like: `/C:/org/file.org` which was then displayed as: `ile.org`.

Some issue was reported in https://github.com/nvim-orgmode/orgmode/issues/911, but not fixed on Windows.