fix(files): remove string tree-sitter parser (#1041)
* fix(files): remove string tree-sitter parser

With the change on nightly https://github.com/neovim/neovim/pull/35988,
string parser creates a scratch buffer under the hood, and uses that as
a source. This drastically decreases the performance of the current
code.

This commit refactors the code to remove the string parser, and adds own
loading of the org buffers for the tree-sitter usage.
Performance should be same as before, with the downside that all of the
org files that are in org_agenda_files are now loaded as unlisted
buffers.

* ci(tests): Run tests only on version 0.11.3+

For some reason, tests on older versions fail without reporting any
actual issues in the tests. Skip running tests on those versions.